[webgpu] 5D, 6D support (#6321)
BUG: #6252

+++ tfjs-backend-webgpu/src/matmul_reduce.ts
@@ -107,18 +107,20 @@ export class MatMulReduceProgram implements WebGPUProgram {
   getUserCode(): string {
     let sampleA;
     if (this.transposeA === false) {
-      sampleA = `return A[batch * batchASize + row * uniforms.dimInner + col];`;
+      sampleA =
+          `return f32(A[batch * batchASize + row * uniforms.dimInner + col]);`;
     } else {
       sampleA =
-          `return A[batch * batchASize + col * uniforms.dimAOuter + row];`;
+          `return f32(A[batch * batchASize + col * uniforms.dimAOuter + row]);`;
     }
 
     let sampleB;
     if (this.transposeB === false) {
       sampleB =
-          `return B[batch * batchBSize + row * uniforms.dimBOuter + col];`;
+          `return f32(B[batch * batchBSize + row * uniforms.dimBOuter + col]);`;
     } else {
-      sampleB = `return B[batch * batchBSize + col * uniforms.dimInner + row];`;
+      sampleB =
+          `return f32(B[batch * batchBSize + col * uniforms.dimInner + row]);`;
     }
 
     let activationSnippet = '', applyActivationSnippet = '';
